fix(shoppinglist-service): update shopping list testing for adding auth restriction (WIP)

+++ src/shoppinglist-service/userShoppingListEntry/default_controller.go
@@ -2,20 +2,29 @@ package userShoppingListEntry
 
 import (
 	"encoding/json"
+	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/lib/router/middleware/auth"
+	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/shoppinglist-service/userShoppingList"
 	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/shoppinglist-service/userShoppingListEntry/model"
 	"net/http"
 	"strconv"
 )
 
-type defaultController struct {
+type DefaultController struct {
 	userShoppingListEntryRepository Repository
+	userShoppingListRepository      userShoppingList.Repository
 }
 
-func NewDefaultController(userShoppingListEntryRepository Repository) *defaultController {
-	return &defaultController{userShoppingListEntryRepository}
+func NewDefaultController(
+	userShoppingListEntryRepository Repository,
+	userShoppingListRepository userShoppingList.Repository,
+) *DefaultController {
+	return &DefaultController{
+		userShoppingListEntryRepository,
+		userShoppingListRepository,
+	}
 }
 
-func (controller defaultController) GetEntries(writer http.ResponseWriter, request *http.Request) {
+func (controller *DefaultController) GetEntries(writer http.ResponseWriter, request *http.Request) {
 	listId, err := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
 
 	if err != nil {
@@ -23,19 +32,34 @@ func (controller defaultController) GetEntries(writer http.ResponseWriter, reque
 		return
 	}
 
-	values, err := controller.userShoppingListEntryRepository.FindAll(listId)
+	list, err := controller.userShoppingListRepository.FindById(listId)
 	if err != nil {
-		http.Error(writer, err.Error(), http.StatusNotFound)
+		http.Error(writer, err.Error(), http.StatusBadRequest)
+		return
 	}
 
-	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(values)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
+	values, err := controller.userShoppingListEntryRepository.FindAll(listId)
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == list.UserId || authUserRole == auth.Administrator {
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusNotFound)
+		}
+
+		writer.Header().Set("Content-Type", "application/json")
+		err = json.NewEncoder(writer).Encode(values)
+
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
 
-func (controller defaultController) GetEntry(writer http.ResponseWriter, request *http.Request) {
+func (controller *DefaultController) GetEntry(writer http.ResponseWriter, request *http.Request) {
 	listId, err := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
 	productId, err := strconv.ParseUint(request.Context().Value("productId").(string), 10, 64)
 
@@ -44,22 +68,36 @@ func (controller defaultController) GetEntry(writer http.ResponseWriter, request
 		return
 	}
 
-	value, err := controller.userShoppingListEntryRepository.FindByIds(listId, productId)
+	list, err := controller.userShoppingListRepository.FindById(listId)
 	if err != nil {
-		if err.Error() == ErrorEntryNotFound {
-			http.Error(writer, err.Error(), http.StatusNotFound)
-			return
-		}
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
+		http.Error(writer, err.Error(), http.StatusBadRequest)
+		return
 	}
 
-	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(value)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
+	value, err := controller.userShoppingListEntryRepository.FindByIds(listId, productId)
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == list.UserId || authUserRole == auth.Administrator {
+		if err != nil {
+			if err.Error() == ErrorEntryNotFound {
+				http.Error(writer, err.Error(), http.StatusNotFound)
+				return
+			}
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+
+		writer.Header().Set("Content-Type", "application/json")
+		err = json.NewEncoder(writer).Encode(value)
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
-func (controller defaultController) PostEntry(writer http.ResponseWriter, request *http.Request) {
+func (controller *DefaultController) PostEntry(writer http.ResponseWriter, request *http.Request) {
 	listId, listIdErr := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
 	productId, productIdErr := strconv.ParseUint(request.Context().Value("productId").(string), 10, 64)
 
@@ -74,20 +112,33 @@ func (controller defaultController) PostEntry(writer http.ResponseWriter, reques
 		return
 	}
 
-	if _, err := controller.userShoppingListEntryRepository.Create(&model.UserShoppingListEntry{
-		ShoppingListId: listId,
-		ProductId:      productId,
-		Count:          requestData.Count,
-		Note:           requestData.Note,
-		Checked:        requestData.Checked,
-	}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
+	list, err := controller.userShoppingListRepository.FindById(listId)
+	if err != nil {
+		http.Error(writer, err.Error(), http.StatusBadRequest)
 		return
 	}
-	writer.WriteHeader(http.StatusCreated)
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == list.UserId || authUserRole == auth.Administrator {
+		if _, err := controller.userShoppingListEntryRepository.Create(&model.UserShoppingListEntry{
+			ShoppingListId: listId,
+			ProductId:      productId,
+			Count:          requestData.Count,
+			Note:           requestData.Note,
+			Checked:        requestData.Checked,
+		}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		writer.WriteHeader(http.StatusCreated)
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
+	}
 }
 
-func (controller defaultController) PutEntry(writer http.ResponseWriter, request *http.Request) {
+func (controller *DefaultController) PutEntry(writer http.ResponseWriter, request *http.Request) {
 	listId, err := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
 	productId, err := strconv.ParseUint(request.Context().Value("productId").(string), 10, 64)
 
@@ -102,19 +153,32 @@ func (controller defaultController) PutEntry(writer http.ResponseWriter, request
 		return
 	}
 
-	if _, err := controller.userShoppingListEntryRepository.Update(&model.UserShoppingListEntry{
-		ShoppingListId: listId,
-		ProductId:      productId,
-		Count:          requestData.Count,
-		Note:           requestData.Note,
-		Checked:        requestData.Checked,
-	}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
+	list, err := controller.userShoppingListRepository.FindById(listId)
+	if err != nil {
+		http.Error(writer, err.Error(), http.StatusBadRequest)
 		return
+	}
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == list.UserId || authUserRole == auth.Administrator {
+		if _, err := controller.userShoppingListEntryRepository.Update(&model.UserShoppingListEntry{
+			ShoppingListId: listId,
+			ProductId:      productId,
+			Count:          requestData.Count,
+			Note:           requestData.Note,
+			Checked:        requestData.Checked,
+		}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
 
-func (controller defaultController) DeleteEntry(writer http.ResponseWriter, request *http.Request) {
+func (controller *DefaultController) DeleteEntry(writer http.ResponseWriter, request *http.Request) {
 	listId, err := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
 	productId, err := strconv.ParseUint(request.Context().Value("productId").(string), 10, 64)
 
@@ -123,8 +187,19 @@ func (controller defaultController) DeleteEntry(writer http.ResponseWriter, requ
 		return
 	}
 
-	if err := controller.userShoppingListEntryRepository.Delete(&model.UserShoppingListEntry{ShoppingListId: listId, ProductId: productId}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
+	list, err := controller.userShoppingListRepository.FindById(listId)
+	if err != nil {
+		http.Error(writer, err.Error(), http.StatusBadRequest)
 		return
+	}
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == list.UserId || authUserRole == auth.Administrator {
+		if err := controller.userShoppingListEntryRepository.Delete(&model.UserShoppingListEntry{ShoppingListId: listId, ProductId: productId}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
 	}
 }
